services/database-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
aiosqlite>=0.20.0
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
python-dotenv>=1.0.1
